fix(payment services): move refresh control to payment services
